* array.c (ary_memcpy): cast to int to suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@42193 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ array.c
@@ -69,7 +69,7 @@ ary_memcpy(VALUE ary, long beg, long argc, const VALUE *argv)
 {
 #if 1
     if (OBJ_PROMOTED(ary)) {
-	if (argc > (128/sizeof(VALUE)) /* is magic number (cache line size) */) {
+	if (argc > (int)(128/sizeof(VALUE)) /* is magic number (cache line size) */) {
 	    rb_gc_writebarrier_remember_promoted(ary);
 	    RARRAY_PTR_USE(ary, ptr, {
 		MEMCPY(ptr+beg, argv, VALUE, argc);
